Renaming namespace Recipe to RecipeManager to avoid naming conflicts with Recipe class
