+++ CIS560-RecipeManager/CIS560-RecipeManager/RecipeManager/Recipe.cs
@@ -11,7 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 using CIS560_RecipeManager;
 
-namespace CIS560_RecipeManager
+namespace CIS560_RecipeManager.RecipeManager
 {
     public class Recipe
     {
