+++ CIS560-RecipeManager/CIS560-RecipeManager/RecipeManager/RecipeCategory.cs
@@ -1,4 +1,4 @@
-﻿namespace CIS560_RecipeManager
+﻿namespace CIS560_RecipeManager.RecipeManager
 {
     public class RecipeCategory
     {
