+++ CIS560-RecipeManager/CIS560-RecipeManager/RecipeManager/RecipeController.cs
@@ -1,10 +1,13 @@
-﻿namespace CIS560_RecipeManager.RecipeManager
+﻿
+
+namespace CIS560_RecipeManager.RecipeManager
 {
     public class RecipeController
     {
         private IQuery QueryRepository;
+        private RecipeInventory recipeInventory;
 
-        public RecipeController(IQuery query)
+        public RecipeController(IQuery query, RecipeInventory recipeInventory)
         {
             QueryRepository = query;
         }
