+++ CIS560-RecipeManager/CIS560-RecipeManager/RecipeManager/uiRecipe.resx
