+++ CIS560-RecipeManager/CIS560-RecipeManager/Repository/QueryStub/QueryStub.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using CIS560_RecipeManager.RecipeManager;
+using System;
 using System.Collections.Generic;
 
 namespace CIS560_RecipeManager.Repository
