+++ CIS560-RecipeManager/CIS560-RecipeManager/Repository/RecipeRepository.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using CIS560_RecipeManager.RecipeManager;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
